Removed overhead of opSpot.applyMultiply
I noticed an overhead time of 10-20% when computing b = A*x, with
A a Spot operator and x a column vector. I suspect that in line 96
the operation x(:,i) and y(:,i) for vectors do some internal
reshaping (or index operations), causing extra overhead. I added
an extra conditional to check if x is column vector, then simply
y = op.multiply(x,mode) is run to circumvent this.